Add pH report for chem feeder via MQTT device
This chem feeder via MQTT makes use of MQTT to provide pH reading
value, active status, and alarm status.

A process or device would publish its pH and status to the MQTT
service which AquaLinkD would subscribe to.
The process or device can acquire the pH and status in any method.
An implementation of this is the pH Dosing using a web camera to
extract the pH and status and publish the data to the MQTT service.

Signed-off-by: Loc Ho <locnhinho@gmail.com>